refactor: separate components from recruit page
