+++ frontend/src/app/recruit/page.tsx
@@ -1,5 +1,10 @@
 "use client";
 
+import {
+  RecruitCard,
+  RecruitStartButton,
+  RecruitTypeButtons
+} from "@/components/recruit/Recruit";
 import axios from "axios";
 import { useCallback, useEffect, useState } from "react";
 
@@ -38,63 +43,6 @@ const recruitFromData = ({
 const defaultState: IRecruitAPIResponse = {
   normal: [],
   final: []
-};
-
-const RecruitButton = ({ label, onClick }: { label: string; onClick: any }) => (
-  <button
-    className="p-2 bg-cyan-400 rounded-md font-bold text-xs"
-    onClick={onClick}
-  >
-    {label}
-  </button>
-);
-
-interface RecruitTypeButtonsProps {
-  recruitTypes: string[];
-  curRecruitType: string;
-  updateFn: (idx: number) => void;
-}
-
-const RecruitTypeButtons = ({
-  recruitTypes,
-  curRecruitType,
-  updateFn
-}: RecruitTypeButtonsProps) => {
-  return (
-    <div className="flex space-x-4">
-      {recruitTypes.map((recruitType, idx) => (
-        <button
-          key={recruitType}
-          className={`border-2 border-gray-400 p-2 rounded-md w-20 h-16 text-xs ${recruitType === curRecruitType ? "bg-yellow-200" : ""}`}
-          onClick={() => updateFn(idx)}
-        >
-          {recruitType === "" ? (
-            <span>상시 모집</span>
-          ) : (
-            <div className="flex flex-col">
-              <span>픽업 모집</span>
-              <span className="text-[10px] font-bold">{recruitType}</span>
-            </div>
-          )}
-        </button>
-      ))}
-    </div>
-  );
-};
-
-const RecruitCard = ({ card }: { card: IRecruit }) => {
-  const matched = card.name.match(/^(.*?)(\s*\(.*\))?$/);
-  const name = matched ? matched[1].trim() : "";
-  const cloth = matched && matched[2] ? matched[2]?.trim() : "";
-  return (
-    <div className={`text-center m-2 border-gray-400 p-2 w-16 h-20`}>
-      <img src={`/FX_TEX_GT_${card.star}.png`} className="w-12 h-12" />
-      <div className="flex flex-col overflow-hidden">
-        <span className="text-xs">{name}</span>
-        {cloth && <span className="text-[8px]">{cloth}</span>}
-      </div>
-    </div>
-  );
 };
 
 const useFetchRecruitData = () => {
@@ -272,16 +220,16 @@ const RecruitPage: React.FC = () => {
 
       {!running && (
         <div className="flex justify-center space-x-4">
-          <RecruitButton onClick={doRecruit} label="모집 하기" />
-          <RecruitButton
+          <RecruitStartButton onClick={doRecruit} label="모집 하기" />
+          <RecruitStartButton
             onClick={() => repeatRecruit(200, true)}
             label="뽑을 때 까지 모집 하기"
           />
-          <RecruitButton
+          <RecruitStartButton
             onClick={() => repeatRecruit(200, false)}
             label="200연 모집 하기"
           />
-          <RecruitButton onClick={resetRecruit} label="초기화" />
+          <RecruitStartButton onClick={resetRecruit} label="초기화" />
         </div>
       )}
       <div>모집 포인트: {recruitPoint}</div>
